bot_fa/pyproject.toml
[tool.poetry]
name = "bot-fa"
version = "0.1.0"
description = ""
authors = ["Данил Воронин <melinbodb@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
opencv-python = "^4.8.1.78"
pillow = "^10.1.0"
mss = "^9.0.1"
numpy = "^1.26.1"
matplotlib = "^3.8.0"
pyautogui = "^0.9.54"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
